client resources and models finished execpt settings

+++ client/models/surveys.py
@@ -1,13 +1,13 @@
 from db import db
 
-# Table surveys in client
+# to store which survey were answered
 class SurveyModel(db.Model):
 
     __tablename__ = "surveys"
     id = db.Column(db.Integer, primary_key=True)
     surveyid = db.Column(db.String(100))
-    longterm = db.Column(db.Integer(1))
-    processed = db.Column(db.Integer(1))
+    longterm = db.Column(db.Integer)
+    processed = db.Column(db.Integer)
 
     def __init__(self, surveyid, longterm, processed):
         self.surveyid = surveyid
